+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 1.2.0 - 2020/01/19
+- Added the ability to include class aliases.
 - Added the `short_array_syntax` argument.
 
 ## 1.1.1 - 2020/11/06